#137 GaussianRBFGrid tested and working on TileCoding.

+++ RLSimion/actor.cpp
@@ -24,7 +24,7 @@ Actor::~Actor() {}
 
 void Actor::deferredLoadStep()
 {
-	unsigned int controllerActionIndex, actorActionIndex;
+	size_t controllerActionIndex, actorActionIndex;
 	size_t numWeights;
 	IMemBuffer *pWeights;
 	State* s= SimionApp::get()->pWorld->getDynamicModel()->getStateInstance();
@@ -32,20 +32,20 @@ void Actor::deferredLoadStep()
 	
 	if (m_pInitController.shared_ptr())
 	{
-		int numActionDims = std::min(m_pInitController->getNumOutputs(), m_policyLearners.size());
+		size_t numActionDims = std::min(m_pInitController->getNumOutputs(), m_policyLearners.size());
 		Logger::logMessage(MessageType::Info, "Initializing the policy weights using the base controller");
 		//initialize the weights using the controller's output at each center point in state space
-		for (int actionIndex = 0; actionIndex < numActionDims; actionIndex++)
+		for (size_t actionIndex = 0; actionIndex < numActionDims; actionIndex++)
 		{
 			controllerActionIndex = m_pInitController->getOutputActionIndex(actionIndex);
-			for (unsigned int actorActionIndex = 0; actorActionIndex < m_policyLearners.size(); actorActionIndex++)
+			for (size_t actorActionIndex = 0; actorActionIndex < m_policyLearners.size(); actorActionIndex++)
 			{
 				if (controllerActionIndex == m_policyLearners[actorActionIndex]->getPolicy()->getOutputActionIndex())
 				{
 					//controller's output action index and actor's match, so we use it to initialize
 					numWeights = m_policyLearners[actorActionIndex]->getPolicy()->getDetPolicyStateVFA()->getNumWeights();
 					pWeights = m_policyLearners[actorActionIndex]->getPolicy()->getDetPolicyStateVFA()->getWeights();
-					for (unsigned int i = 0; i < numWeights; i++)
+					for (size_t i = 0; i < numWeights; i++)
 					{
 						m_policyLearners[actorActionIndex]->getPolicy()->getDetPolicyStateVFA()->getStateFeatureMap()->getFeatureStateAction(i, s, a);
 						m_pInitController->selectAction(s, a);
