+++ RLSimion/controller-adaptive.h
@@ -7,7 +7,7 @@
 
 class ExtendedWindTurbineVidalController: public WindTurbineVidalController
 {
-	unsigned int m_aActionId, m_KAlphaActionId, m_KPActionId;
+	size_t m_aActionId, m_KAlphaActionId, m_KPActionId;
 
 	CHILD_OBJECT_FACTORY<ICritic> m_pCritic;
 	CHILD_OBJECT_FACTORY<PolicyLearner> m_pALearner, m_pKAlphaLearner, m_pKPLearner;
@@ -22,7 +22,7 @@ public:
 
 class ExtendedWindTurbineBoukhezzarController : public WindTurbineBoukhezzarController
 {
-	unsigned int m_C0ActionId, m_KPActionId;
+	size_t m_C0ActionId, m_KPActionId;
 	CHILD_OBJECT_FACTORY<ICritic> m_pCritic;
 	CHILD_OBJECT_FACTORY<PolicyLearner> m_pC0Learner, m_pKPLearner;
 
@@ -37,7 +37,7 @@ public:
 class ExtendedWindTurbineJonkmanController : public WindTurbineJonkmanController
 {
 	double m_td;
-	unsigned int m_KPActionId;
+	size_t m_KPActionId;
 	CHILD_OBJECT_FACTORY<ICritic> m_pCritic;
 	CHILD_OBJECT_FACTORY<PolicyLearner> m_pPC_KP_Learner;
