+++ RLSimion/worlds/aux-rewards.h
@@ -12,7 +12,7 @@ class DistanceReward2D : public IRewardComponent
 	double m_maxDist= 1.0;
 public:
 	DistanceReward2D(Descriptor& stateDescr, const char* var1xName, const char* var1yName, const char* var2xName, const char* var2yName);
-	DistanceReward2D(Descriptor& stateDescr, int var1xId, int var1yId, int var2xId, int var2yId);
+	DistanceReward2D(Descriptor& stateDescr, size_t var1xId, size_t var1yId, size_t var2xId, size_t var2yId);
 	double getReward(const State *s, const Action *a, const State *s_p);
 	const char* getName() { return "reward"; }
 	double getMin();
