+++ RLSimion/worlds/pull-box-1.h
@@ -10,14 +10,14 @@ class PullBox1 : public DynamicModel
 {
 
 	/// State variables
-	int m_target_X, m_target_Y;
-	int m_state_X, m_state_Y;
-	int m_rob1_X, m_rob1_Y;
-	int m_box_X, m_box_Y;
-	int m_D_Br1X, m_D_Br1Y;;
-	int m_D_BtX, m_D_BtY;
-	int m_theta_r1;
-	int m_boxTheta;
+	size_t m_target_X, m_target_Y;
+	size_t m_state_X, m_state_Y;
+	size_t m_rob1_X, m_rob1_Y;
+	size_t m_box_X, m_box_Y;
+	size_t m_D_Br1X, m_D_Br1Y;;
+	size_t m_D_BtX, m_D_BtY;
+	size_t m_theta_r1;
+	size_t m_boxTheta;
 
 	BulletPhysics* m_pBulletPhysics;
