+++ RLSimion/worlds/push-box-1.h
@@ -8,13 +8,13 @@ class BulletPhysics;
 class PushBox1 : public DynamicModel
 {
 	/// State variables
-	int m_target_X, m_target_Y;
-	int m_rob1_X, m_rob1_Y;
-	int m_box_X, m_box_Y;
-	int m_D_BrX, m_D_BrY;
-	int m_D_BtX, m_D_BtY;
-	int m_theta;
-	int m_boxTheta;
+	size_t m_target_X, m_target_Y;
+	size_t m_rob1_X, m_rob1_Y;
+	size_t m_box_X, m_box_Y;
+	size_t m_D_BrX, m_D_BrY;
+	size_t m_D_BtX, m_D_BtY;
+	size_t m_theta;
+	size_t m_boxTheta;
 
 	///Graphic initialization
 	BulletPhysics* m_pBulletPhysics;
